+++ lib/hive.dart
@@ -6,20 +6,22 @@ import 'dart:async';
 import 'dart:convert';
 import 'dart:typed_data';
 
+import 'package:hive/src/box/default_compaction_strategy.dart';
+import 'package:hive/src/box/default_key_comparator.dart';
 import 'package:hive/src/hive_impl.dart';
 import 'package:meta/meta.dart';
-
-export 'src/hive_object.dart';
 
 part 'src/annotations/hive_field.dart';
 part 'src/annotations/hive_type.dart';
 part 'src/binary/binary_reader.dart';
 part 'src/binary/binary_writer.dart';
+part 'src/box/box_base.dart';
 part 'src/box/box.dart';
 part 'src/box/lazy_box.dart';
 part 'src/registry/type_registry.dart';
 part 'src/registry/type_adapter.dart';
 part 'src/hive_error.dart';
+part 'src/hive_object.dart';
 part 'src/hive.dart';
 
 /// Global constant to access Hive.