+++ lib/src/backend/storage_backend.dart
@@ -1,17 +1,18 @@
 import 'package:hive/hive.dart';
 import 'package:hive/src/binary/frame.dart';
 import 'package:hive/src/box/keystore.dart';
+import 'package:hive/src/crypto_helper.dart';
 
-export 'package:hive/src/backend/storage_backend_stub.dart'
-    if (dart.library.io) 'package:hive/src/backend/storage_backend_vm.dart'
-    if (dart.library.html) 'package:hive/src/backend/storage_backend_js.dart';
+export 'package:hive/src/backend/stub/backend_manager.dart'
+    if (dart.library.io) 'package:hive/src/backend/vm/backend_manager.dart'
+    if (dart.library.html) 'package:hive/src/backend/js/backend_manager.dart';
 
 abstract class StorageBackend {
   String get path;
 
   bool get supportsCompaction;
 
-  Future<void> initialize(TypeRegistry registry, Keystore keystore);
+  Future<void> initialize(TypeRegistry registry, Keystore keystore, bool lazy);
 
   Future<dynamic> readValue(Frame frame);
 
@@ -24,4 +25,11 @@ abstract class StorageBackend {
   Future<void> close();
 
   Future<void> deleteFromDisk();
+}
+
+abstract class BackendManagerInterface {
+  Future<StorageBackend> open(
+      String name, String path, bool crashRecovery, CryptoHelper crypto);
+
+  Future<void> deleteBox(String name, String path);
 }