+++ lib/src/box/box_impl.dart
@@ -3,10 +3,10 @@ import 'dart:async';
 import 'package:hive/hive.dart';
 import 'package:hive/src/backend/storage_backend.dart';
 import 'package:hive/src/binary/frame.dart';
-import 'package:hive/src/box/box_base.dart';
+import 'package:hive/src/box/box_base_impl.dart';
 import 'package:hive/src/hive_impl.dart';
 
-class BoxImpl<E> extends BoxBase<E> {
+class BoxImpl<E> extends BoxBaseImpl<E> implements Box<E> {
   BoxImpl(
     HiveImpl hive,
     String name,