+++ lib/src/box/default_key_comparator.dart
@@ -1,0 +1,7 @@
+int defaultKeyComparator(dynamic k1, dynamic k2) {
+  if (k1.runtimeType == k2.runtimeType) {
+    return (k1 as Comparable).compareTo(k2);
+  } else {
+    return k1 is int ? -1 : 1;
+  }
+}