+++ lib/src/hive.dart
@@ -2,12 +2,6 @@ part of hive;
 
 /// The main API interface of Hive. Available through the `Hive` constant.
 abstract class HiveInterface implements TypeRegistry {
-  /// The home directory of Hive.
-  ///
-  /// All box files will be stored in this directory. In the browser, this is
-  /// always `null`.
-  String get path;
-
   /// Initialize Hive by giving it a home directory.
   ///
   /// (Not necessary in the browser)
@@ -20,30 +14,39 @@ abstract class HiveInterface implements TypeRegistry {
   Future<Box<E>> openBox<E>(
     String name, {
     List<int> encryptionKey,
-    KeyComparator keyComparator,
-    CompactionStrategy compactionStrategy,
+    KeyComparator keyComparator = defaultKeyComparator,
+    CompactionStrategy compactionStrategy = defaultCompactionStrategy,
     bool crashRecovery = true,
-    bool lazy = false,
+    String path,
+    Uint8List bytes,
   });
 
-  /// Opens an in-memory box from a list of bytes. It does not persist changes.
+  /// Opens a lazy box.
   ///
-  /// This is useful for opening boxes from an asset file.
-  Future<Box<E>> openMemoryBox<E>(
+  /// If the box is already open, the instance is returned and all provided
+  /// parameters are being ignored.
+  Future<LazyBox<E>> openLazyBox<E>(
     String name, {
-    Uint8List bytes,
     List<int> encryptionKey,
-    KeyComparator keyComparator,
+    KeyComparator keyComparator = defaultKeyComparator,
+    CompactionStrategy compactionStrategy = defaultCompactionStrategy,
+    bool crashRecovery = true,
+    String path,
   });
 
   /// Returns a previously opened box.
   Box<E> box<E>(String name);
+
+  /// Returns a previously opened lazy box.
+  LazyBox<E> lazyBox<E>(String name);
 
   /// Checks if a specific box is currently open.
   bool isBoxOpen(String name);
 
   /// Closes all open boxes.
   Future<void> close();
+
+  Future<void> deleteBoxFromDisk(String nme);
 
   /// Deletes all currently open boxes from disk.
   ///